@@ -121,7 +121,7 @@
     "tsickle": "^0.34.0",
     "tslib": "^1.9.3",
     "tslint": "~5.11.0",
-    "typescript": "~3.1.6",
+    "typescript": "~3.2.2",
     "webpack-cli": "^3.1.2"
   }
 }
