@@ -1,8 +1,13 @@
 import { InjectionToken, Provider } from '@angular/core';
+
 import { ActionReducerMap, MetaReducer, ActionReducer } from '@ngrx/store';
+
 import { CartsState, CartState, CART_DATA } from './../cart-state';
-import { reducer as cartReducer } from './cart.reducer';
+import { LOGOUT } from '../../../auth/store/actions/login-logout.action';
+import { PLACE_ORDER_SUCCESS } from '../../../checkout/store/actions/checkout.action';
 import { loaderReducer } from '../../../state/utils/loader/loader.reducer';
+
+import { reducer as cartReducer } from './cart.reducer';
 
 export function getReducers(): ActionReducerMap<CartsState> {
   return {
@@ -23,10 +28,7 @@ export function clearCartState(
   reducer: ActionReducer<any>
 ): ActionReducer<any> {
   return function(state, action) {
-    if (
-      action.type === '[Auth] Logout' ||
-      action.type === '[Checkout] Place Order Success'
-    ) {
+    if (action.type === LOGOUT || action.type === PLACE_ORDER_SUCCESS) {
       state = undefined;
     }
     return reducer(state, action);
