@@ -1,6 +1,7 @@
 import { StaticProvider } from '@angular/core';
-import { OccConfig } from '../../occ/config/occ-config';
+
 import { AuthConfig } from '../../auth/config/auth-config';
+import { OccConfig } from '../../occ/config/occ-config';
 
 export interface StandardCmsComponentConfig {
   SiteContextSelectorComponent?: CmsComponentMapping;
@@ -25,6 +26,7 @@ export interface JspIncludeCmsComponentConfig {
 }
 
 export const JSP_INCLUDE_CMS_COMPONENT_TYPE = 'JspIncludeComponent';
+export const FLEX_CMS_COMPONENT_TYPE = 'FlexCmsComponent';
 
 export interface CmsComponentMapping {
   selector?: string;
@@ -42,11 +44,6 @@ export abstract class CmsConfig extends OccConfig implements AuthConfig {
   authentication?: {
     client_id?: string;
     client_secret?: string;
-  };
-
-  defaultPageIdForType?: {
-    ProductPage?: string[];
-    CategoryPage?: string[];
   };
 
   cmsComponents?: CMSComponentConfig;
