@@ -1,13 +1,26 @@
 import { Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
-import * as fromStore from '../store';
-import { filter, tap, map, take, withLatestFrom } from 'rxjs/operators';
+
 import { select, Store } from '@ngrx/store';
-import { Page } from '../model/page.model';
+
+import { Observable } from 'rxjs';
+import {
+  filter,
+  tap,
+  map,
+  withLatestFrom,
+  switchMap,
+  take
+} from 'rxjs/operators';
+
+import * as fromStore from '../store';
+import { LoaderState } from '../../state';
 import { ContentSlotData } from '../model/content-slot-data.model';
-import { DefaultPageService } from '../services/default-page.service';
+import { NodeItem } from '../model/node-item.model';
+import { Page } from '../model/page.model';
 import { StateWithCms } from '../store/cms-state';
 import { CmsComponent } from '../../occ/occ-models/cms-component.models';
+import { RoutingService } from '../../routing/facade/routing.service';
+import { PageContext } from '../../routing/models/page-context.model';
 
 @Injectable({
   providedIn: 'root'
@@ -17,13 +30,13 @@ export class CmsService {
 
   constructor(
     private store: Store<StateWithCms>,
-    private defaultPageService: DefaultPageService
+    private routingService: RoutingService
   ) {}
 
   /**
    * Set _launchInSmartEdit value
    */
-  set launchInSmartEdit(value) {
+  set launchInSmartEdit(value: boolean) {
     this._launchInSmartEdit = value;
   }
 
@@ -38,7 +51,13 @@ export class CmsService {
    * Get current CMS page data
    */
   getCurrentPage(): Observable<Page> {
-    return this.store.pipe(select(fromStore.getLatestPage));
+    return this.routingService
+      .getPageContext()
+      .pipe(
+        switchMap(pageContext =>
+          this.store.select(fromStore.getPageData(pageContext))
+        )
+      );
   }
 
   /**
@@ -68,9 +87,13 @@ export class CmsService {
    * @param position : content slot position
    */
   getContentSlot(position: string): Observable<ContentSlotData> {
-    return this.store.pipe(
-      select(fromStore.currentSlotSelectorFactory(position)),
-      filter(Boolean)
+    return this.routingService.getPageContext().pipe(
+      switchMap(pageContext =>
+        this.store.pipe(
+          select(fromStore.currentSlotSelectorFactory(pageContext, position)),
+          filter(Boolean)
+        )
+      )
     );
   }
 
@@ -78,7 +101,7 @@ export class CmsService {
    * Given navigation node uid, get items (with id and type) inside the navigation entries
    * @param navigationNodeUid : uid of the navigation node
    */
-  getNavigationEntryItems(navigationNodeUid: string): Observable<any> {
+  getNavigationEntryItems(navigationNodeUid: string): Observable<NodeItem> {
     return this.store.pipe(
       select(fromStore.itemsSelectorFactory(navigationNodeUid))
     );
@@ -92,7 +115,7 @@ export class CmsService {
   loadNavigationItems(
     rootUid: string,
     itemList: { id: string; superType: string }[]
-  ) {
+  ): void {
     this.store.dispatch(
       new fromStore.LoadNavigationItems({
         nodeId: rootUid,
@@ -104,15 +127,20 @@ export class CmsService {
   /**
    * Refresh the content of the latest cms page
    */
-  refreshLatestPage() {
-    this.store.dispatch(new fromStore.RefreshLatestPage());
+  refreshLatestPage(): void {
+    this.routingService
+      .getPageContext()
+      .pipe(take(1))
+      .subscribe(pageContext =>
+        this.store.dispatch(new fromStore.LoadPageData(pageContext))
+      );
   }
 
   /**
    * Refresh cms component's content
    * @param uid : component uid
    */
-  refreshComponent(uid: string) {
+  refreshComponent(uid: string): void {
     this.store.dispatch(new fromStore.LoadComponent(uid));
   }
 
@@ -120,45 +148,17 @@ export class CmsService {
    * Given pageContext, return whether the CMS page data exists or not
    * @param pageContext
    */
-  hasPage(pageContext): Observable<boolean> {
-    let tryTimes = 0;
-
+  hasPage(pageContext: PageContext): Observable<boolean> {
     return this.store.pipe(
-      select(fromStore.getPageEntities),
-      map((entities: { [key: string]: Page }) => {
-        let key = pageContext.id + '_' + pageContext.type;
-        let found = !!entities[key];
-        if (!found) {
-          const defaultPageIds = this.defaultPageService.getDefaultPageIdsBytype(
-            pageContext.type
-          );
-          if (defaultPageIds) {
-            for (let i = 0, len = defaultPageIds.length; i < len; i++) {
-              key = defaultPageIds[i] + '_' + pageContext.type;
-              found =
-                entities[key] &&
-                entities[key].seen.indexOf(pageContext.id) > -1;
-              if (found) {
-                break;
-              }
-            }
-          }
-        }
-        // found page directly from store
-        if (found && tryTimes === 0) {
-          this.store.dispatch(new fromStore.UpdateLatestPageKey(key));
-        }
-        return found;
-      }),
-      tap(found => {
-        // if not found, load this cms page
-        if (!found) {
-          tryTimes = tryTimes + 1;
+      select(fromStore.getIndexEntity(pageContext)),
+      tap((entity: LoaderState<string>) => {
+        const attemptedLoad = entity.loading || entity.success || entity.error;
+        if (!attemptedLoad) {
           this.store.dispatch(new fromStore.LoadPageData(pageContext));
         }
       }),
-      filter(found => found || tryTimes === 3),
-      take(1)
+      filter(entity => entity.success || entity.error),
+      map(entity => entity.success)
     );
   }
 }
