@@ -1,8 +1,11 @@
 export * from './config/cms-config';
+export { defaultCmsModuleConfig } from './config/default-cms-config';
+
 export * from './model/page.model';
 export * from './model/node-item.model';
 export * from './model/content-slot-data.model';
 export * from './model/content-slot-component-data.model';
+
 export * from './occ/index';
 export * from './occ/cms-occ.module';
 
@@ -12,4 +15,8 @@ export * from './store/selectors/index';
 
 export * from './facade/index';
 export * from './cms.module';
-export { defaultCmsModuleConfig } from './config/default-cms-config';
+
+export * from './services/index';
+
+export * from './page/index';
+export * from './page/page.module';
