@@ -5,9 +5,9 @@ export interface Page {
   pageId?: string;
   catalogUuid?: string;
   name?: string;
+  type?: string;
   title?: string;
   template?: string;
   loadTime?: number;
-  seen?: Array<string>;
   slots: { [key: string]: ContentSlotData };
 }
