@@ -1,19 +1,21 @@
 import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
-import { hot, cold } from 'jasmine-marbles';
-import { Observable, of } from 'rxjs';
-
-import { OccCmsService } from '../../occ/occ-cms.service';
-import { CmsConfig } from '../../config/cms-config';
-import * as fromEffects from './component.effect';
-import * as fromActions from '../actions/component.action';
-
 import { provideMockActions } from '@ngrx/effects/testing';
 import { StoreModule } from '@ngrx/store';
-import { PageType, CmsComponent } from '../../../occ/occ-models/index';
-import { RoutingService } from '../../../routing/index';
+
+import { Observable, of } from 'rxjs';
+
+import { hot, cold } from 'jasmine-marbles';
+
+import * as fromActions from '../actions/component.action';
+import { CmsConfig } from '../../config/cms-config';
 import { defaultCmsModuleConfig } from '../../config/default-cms-config';
+import { OccCmsService } from '../../occ/occ-cms.service';
+import { RoutingService } from '../../../routing/index';
+import { PageType, CmsComponent } from '../../../occ/occ-models/index';
+
+import * as fromEffects from './component.effect';
 
 const router = {
   state: {
@@ -26,7 +28,7 @@ const router = {
 };
 
 class MockRoutingService {
-  getRouterState() {
+  getRouterState(): Observable<any> {
     return of(router);
   }
 }
