@@ -1,53 +1,65 @@
 import { TestBed } from '@angular/core/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
-import { hot, cold } from 'jasmine-marbles';
+import { provideMockActions } from '@ngrx/effects/testing';
+import { StoreModule, Action } from '@ngrx/store';
+
 import { Observable, of } from 'rxjs';
 
-import { OccCmsService } from '../../occ/occ-cms.service';
-import { DefaultPageService } from '../../services/default-page.service';
-import { CmsConfig } from '../../config/cms-config';
-import * as fromEffects from './page.effect';
+import { hot, cold } from 'jasmine-marbles';
+
 import * as fromActions from '../actions';
-import { Page } from '../../model/page.model';
-import { PageType } from '../../../occ/occ-models';
-import { provideMockActions } from '@ngrx/effects/testing';
-import { StoreModule } from '@ngrx/store';
-import * as fromCmsReducer from '../../../cms/store/reducers';
+import { Logout, Login } from '../../../auth';
 import { PageContext, RoutingService } from '../../../routing';
+import { LanguageChange } from '../../../site-context';
+import { CmsConfig } from '../../config/cms-config';
 import { defaultCmsModuleConfig } from '../../config/default-cms-config';
+import { Page } from '../../model/page.model';
+import { OccCmsService } from '../../occ/occ-cms.service';
+import { PageType, CmsComponent } from '../../../occ/occ-models';
+import * as fromCmsReducer from '../../../cms/store/reducers';
+
+import * as fromEffects from './page.effect';
 
 export function mockDateNow(): number {
   return 1000000000000;
 }
 
+const context: PageContext = {
+  id: 'homepage',
+  type: PageType.CONTENT_PAGE
+};
 const mockRouterState = {
   state: {
     cmsRequired: true,
-    context: {}
+    context
   }
 };
 
-const mockRoutingService = {
-  routerState$: of(mockRouterState)
-};
+class RoutingServiceMock {
+  getRouterState(): Observable<any> {
+    return of(mockRouterState);
+  }
+}
 
 describe('Page Effects', () => {
-  let actions$: Observable<any>;
+  let actions$: Observable<Action>;
   let occService: OccCmsService;
-  let defaultPageService: DefaultPageService;
   let effects: fromEffects.PageEffects;
+  let routingService: RoutingService;
 
   const cmsComponentData: any[] = [
     {
       uid: 'comp1',
       typeCode: 'SimpleBannerComponent',
-      uuid: 'compUuid1'
+      uuid: 'compUuid1',
+      flextype: undefined
     },
     {
       uid: 'comp2',
-      typeCode: 'CMSLinkComponent',
-      uuid: 'compUuid2'
+      typeCode: 'FlexCmsComponent',
+      uuid: 'compUuid2',
+      flextype: 'AccountAddressBookComponent'
     }
   ];
 
@@ -57,6 +69,7 @@ describe('Page Effects', () => {
     name: 'testPage',
     template: 'testTemplate',
     title: 'testPageTitle',
+    typeCode: 'ContentPage',
     contentSlots: {
       contentSlot: [
         {
@@ -87,13 +100,15 @@ describe('Page Effects', () => {
       uid: 'comp1',
       typeCode: 'SimpleBannerComponent',
       uuid: 'compUuid1',
-      catalogUuid: undefined
+      catalogUuid: undefined,
+      flextype: undefined
     },
     {
       uid: 'comp2',
-      typeCode: 'CMSLinkComponent',
+      typeCode: 'FlexCmsComponent',
       uuid: 'compUuid2',
-      catalogUuid: undefined
+      catalogUuid: undefined,
+      flextype: 'AccountAddressBookComponent'
     }
   ];
 
@@ -102,9 +117,9 @@ describe('Page Effects', () => {
     loadTime: 1000000000000,
     name: 'testPage',
     pageId: 'testPageId',
+    type: 'ContentPage',
     template: 'testTemplate',
     title: 'testPageTitle',
-    seen: new Array<string>(),
     catalogUuid: 'mockPageCatalogUuid',
     slots: {
       testPosition: {
@@ -125,98 +140,109 @@ describe('Page Effects', () => {
       ],
       providers: [
         OccCmsService,
-        { provide: RoutingService, useValue: mockRoutingService },
+        { provide: RoutingService, useClass: RoutingServiceMock },
         { provide: CmsConfig, useValue: defaultCmsModuleConfig },
-        DefaultPageService,
         fromEffects.PageEffects,
         provideMockActions(() => actions$)
       ]
     });
 
     occService = TestBed.get(OccCmsService);
-    defaultPageService = TestBed.get(DefaultPageService);
     effects = TestBed.get(fromEffects.PageEffects);
+    routingService = TestBed.get(RoutingService);
     Date.now = mockDateNow;
 
     spyOn(occService, 'loadPageData').and.returnValue(of(cmsPageData));
-    spyOn(defaultPageService, 'getDefaultPageIdsBytype').and.returnValue([
-      'productList'
-    ]);
   });
 
-  describe('loadPage$', () => {
-    it('should emit actions LoadPageDataSuccess and GetComponentFromPage for ContentPage type', () => {
-      const context: PageContext = {
-        id: 'testPagId',
-        type: PageType.CONTENT_PAGE
-      };
+  describe('refreshPage$', () => {
+    describe('when a language changes', () => {
+      it('should dispatch LoadPageIndex action', () => {
+        spyOn(routingService, 'getRouterState').and.returnValue(
+          of(mockRouterState)
+        );
 
-      const action = new fromActions.LoadPageData(context);
+        const action = new LanguageChange();
+        const completion = new fromActions.LoadPageData(context);
 
-      page.seen.push(context.id);
-      const pageKey = page.pageId + '_' + context.type;
-      const payload = { key: pageKey, value: page };
+        actions$ = hot('-a', { a: action });
+        const expected = cold('-b', { b: completion });
 
-      const completion1 = new fromActions.LoadPageDataSuccess(payload);
-      const completion2 = new fromActions.GetComponentFromPage(
-        cmsComponentData
-      );
-
-      actions$ = hot('-a', { a: action });
-      const expected = cold('-(bc)', { b: completion1, c: completion2 });
-
-      expect(effects.loadPage$).toBeObservable(expected);
+        expect(effects.refreshPage$).toBeObservable(expected);
+      });
     });
+    describe('when a user logs in', () => {
+      it('should dispatch LoadPageIndex action', () => {
+        spyOn(routingService, 'getRouterState').and.returnValue(
+          of(mockRouterState)
+        );
 
-    it('should emit actions LoadPageDataSuccess and GetComponentFromPage for non-ContentPage type (specific)', () => {
-      const context: PageContext = {
-        id: '1234',
-        type: PageType.PRODUCT_PAGE
-      };
+        const action = new Logout();
+        const completion = new fromActions.LoadPageData(context);
 
-      const action = new fromActions.LoadPageData(context);
+        actions$ = hot('-a', { a: action });
+        const expected = cold('-b', { b: completion });
 
-      page.seen = new Array<string>();
-      page.seen.push(context.id);
-      const pageKey = context.id + '_' + context.type;
-      const payload = { key: pageKey, value: page };
-
-      const completion1 = new fromActions.LoadPageDataSuccess(payload);
-      const completion2 = new fromActions.GetComponentFromPage(
-        cmsComponentData
-      );
-
-      actions$ = hot('-a', { a: action });
-      const expected = cold('-(bc)', { b: completion1, c: completion2 });
-
-      expect(effects.loadPage$).toBeObservable(expected);
+        expect(effects.refreshPage$).toBeObservable(expected);
+      });
     });
+    describe('when a user logs out', () => {
+      it('should dispatch LoadPageIndex action', () => {
+        spyOn(routingService, 'getRouterState').and.returnValue(
+          of(mockRouterState)
+        );
 
-    it('should emit actions LoadPageDataSuccess and GetComponentFromPage for non-ContentPage type (default)', () => {
-      const context: PageContext = {
-        id: '1234',
-        type: PageType.PRODUCT_PAGE
-      };
+        const action = new Login();
+        const completion = new fromActions.LoadPageData(context);
 
-      const action = new fromActions.LoadPageData(context);
+        actions$ = hot('-a', { a: action });
+        const expected = cold('-b', { b: completion });
 
-      cmsPageData.uid = 'productList';
-      page.pageId = 'productList';
-      page.seen = new Array<string>();
-      page.seen.push(context.id);
+        expect(effects.refreshPage$).toBeObservable(expected);
+      });
+    });
+  });
 
-      const pageKey = page.pageId + '_' + context.type;
-      const payload = { key: pageKey, value: page };
+  describe('loadPageData$', () => {
+    describe('when LoadPageData is dispatched', () => {
+      it('should dispatch LoadPageDataSuccess and GetComponentFromPage actions', () => {
+        const mockedComponents: CmsComponent[] = [{ name: 'aComponent' }];
+        spyOn<any>(effects, 'getComponents').and.returnValue(mockedComponents);
 
-      const completion1 = new fromActions.LoadPageDataSuccess(payload);
-      const completion2 = new fromActions.GetComponentFromPage(
-        cmsComponentData
-      );
+        const action = new fromActions.LoadPageData(context);
 
-      actions$ = hot('-a', { a: action });
-      const expected = cold('-(bc)', { b: completion1, c: completion2 });
+        const completion1 = new fromActions.LoadPageDataSuccess(context, page);
+        const completion2 = new fromActions.GetComponentFromPage(
+          mockedComponents
+        );
 
-      expect(effects.loadPage$).toBeObservable(expected);
+        actions$ = hot('-a', { a: action });
+        const expected = cold('-(bc)', {
+          b: completion1,
+          c: completion2
+        });
+
+        expect(effects.loadPageData$).toBeObservable(expected);
+      });
+
+      it('should dispatch LoadPageDataFail action', () => {
+        const error = 'error';
+        spyOn<any>(effects, 'getPageData').and.throwError(error);
+
+        const action = new fromActions.LoadPageData(context);
+
+        const completion = new fromActions.LoadPageDataFail(
+          context,
+          new Error(error)
+        );
+
+        actions$ = hot('-a', { a: action });
+        const expected = cold('-b', {
+          b: completion
+        });
+
+        expect(effects.loadPageData$).toBeObservable(expected);
+      });
     });
   });
 });
