@@ -1,14 +1,16 @@
 import { TestBed } from '@angular/core/testing';
+
 import { StoreModule, Store, select } from '@ngrx/store';
 
-import * as fromReducers from '../reducers/index';
+import { StateWithCms, IndexType } from '../cms-state';
 import * as fromActions from '../actions/index';
+import * as fromReducers from '../reducers/index';
 import * as fromSelectors from '../selectors/page.selectors';
-
-import { Page } from '../../model/page.model';
-import { CmsComponent } from '../../../occ/occ-models/index';
-import { StateWithCms } from '../cms-state';
+import { EntityLoaderState, LoaderState } from '../../../state';
 import { ContentSlotData } from '../../model/content-slot-data.model';
+import { Page } from '../../model/page.model';
+import { CmsComponent, PageType } from '../../../occ/occ-models/index';
+import { PageContext } from '../../../routing/models/page-context.model';
 
 describe('Cms PageData Selectors', () => {
   let store: Store<StateWithCms>;
@@ -19,15 +21,14 @@ describe('Cms PageData Selectors', () => {
     { uid: 'comp3', typeCode: 'NavigationComponent' }
   ];
   const page: Page = {
-    pageId: 'testPageId',
-    name: 'testPage',
-    seen: [],
+    pageId: 'homepage',
+    name: 'HomePage',
     slots: { left: { components } }
   };
-  const payload = { key: 'test', value: page };
 
-  const entities = {
-    test: page
+  const pageContext: PageContext = {
+    id: 'homepage',
+    type: PageType.CONTENT_PAGE
   };
 
   beforeEach(() => {
@@ -41,67 +42,127 @@ describe('Cms PageData Selectors', () => {
     spyOn(store, 'dispatch').and.callThrough();
   });
 
-  describe('getPageEntities', () => {
-    it('should return pages as entities', () => {
-      let result: { [key: string]: Page };
+  describe('getPageStateIndex', () => {
+    it('should return the index part of the state', () => {
+      store.dispatch(new fromActions.LoadPageDataSuccess(pageContext, page));
 
+      let result: IndexType;
       store
-        .pipe(select(fromSelectors.getPageEntities))
-        .subscribe(value => (result = value));
+        .pipe(select(fromSelectors.getPageStateIndex))
+        .subscribe(value => (result = value))
+        .unsubscribe();
+
+      const expectedResult: IndexType = {
+        content: {
+          entities: {
+            homepage: {
+              loading: false,
+              error: false,
+              success: true,
+              value: page.pageId
+            }
+          }
+        },
+        product: { entities: {} },
+        category: { entities: {} },
+        catalog: { entities: {} }
+      };
+
+      expect(result).toEqual(expectedResult);
+    });
+  });
+
+  describe('getIndex', () => {
+    it('should return an index', () => {
+      store.dispatch(new fromActions.LoadPageDataSuccess(pageContext, page));
+
+      let result: EntityLoaderState<string>;
+      store
+        .pipe(select(fromSelectors.getIndex(pageContext)))
+        .subscribe(value => (result = value))
+        .unsubscribe();
+
+      expect(result).toEqual({
+        entities: {
+          homepage: {
+            loading: false,
+            error: false,
+            success: true,
+            value: page.pageId
+          }
+        }
+      });
+    });
+  });
+
+  describe('getIndexEntity', () => {
+    it('should retrn an empty object when there is no entity', () => {
+      let result: LoaderState<string>;
+      store
+        .pipe(select(fromSelectors.getIndexEntity(pageContext)))
+        .subscribe(value => (result = value))
+        .unsubscribe();
 
       expect(result).toEqual({});
-
-      store.dispatch(new fromActions.LoadPageDataSuccess(payload));
-
-      expect(result).toEqual(entities);
     });
-  });
 
-  describe('getLatestPageKey', () => {
-    it('should return the latest page key', () => {
-      let result: string;
+    it('should return an entity from an index', () => {
+      store.dispatch(new fromActions.LoadPageDataSuccess(pageContext, page));
 
+      let result: LoaderState<string>;
       store
-        .pipe(select(fromSelectors.getLatestPageKey))
-        .subscribe(value => (result = value));
+        .pipe(select(fromSelectors.getIndexEntity(pageContext)))
+        .subscribe(value => (result = value))
+        .unsubscribe();
 
-      expect(result).toEqual('');
-
-      store.dispatch(new fromActions.UpdateLatestPageKey(payload.key));
-
-      expect(result).toEqual('test');
+      expect(result).toEqual({
+        loading: false,
+        error: false,
+        success: true,
+        value: page.pageId
+      });
     });
   });
 
-  describe('getLatestPage', () => {
-    it('should return the latest page', () => {
+  describe('getPageEntities', () => {
+    it('should return the entities', () => {
+      store.dispatch(new fromActions.LoadPageDataSuccess(pageContext, page));
+
+      let result: { [id: string]: Page };
+      store
+        .pipe(select(fromSelectors.getPageEntities))
+        .subscribe(value => (result = value))
+        .unsubscribe();
+
+      expect(result).toEqual({ homepage: page });
+    });
+  });
+
+  describe('getPageData', () => {
+    it('should return the page', () => {
+      store.dispatch(new fromActions.LoadPageDataSuccess(pageContext, page));
+
       let result: Page;
-
       store
-        .pipe(select(fromSelectors.getLatestPage))
-        .subscribe(value => (result = value));
+        .pipe(select(fromSelectors.getPageData(pageContext)))
+        .subscribe(value => (result = value))
+        .unsubscribe();
 
-      expect(result).toEqual(undefined);
-
-      store.dispatch(new fromActions.LoadPageDataSuccess(payload));
-      store.dispatch(new fromActions.UpdateLatestPageKey(payload.key));
-
-      expect(result).toEqual(payload.value);
+      expect(result).toEqual(page);
     });
   });
 
   describe('currentSlotSelectorFactory', () => {
     it('should return current slot by position', () => {
+      store.dispatch(new fromActions.LoadPageDataSuccess(pageContext, page));
+
       let result: ContentSlotData;
-
       store
-        .pipe(select(fromSelectors.currentSlotSelectorFactory('left')))
-        .subscribe(value => (result = value));
-
-      expect(result).toEqual(undefined);
-
-      store.dispatch(new fromActions.LoadPageDataSuccess(payload));
-      store.dispatch(new fromActions.UpdateLatestPageKey(payload.key));
+        .pipe(
+          select(fromSelectors.currentSlotSelectorFactory(pageContext, 'left'))
+        )
+        .subscribe(value => (result = value))
+        .unsubscribe();
 
       expect(result).toEqual({ components });
     });
