@@ -11,15 +11,53 @@ export function defaultSiteContextConfigFactory(): SiteContextConfig {
         [LANGUAGE_CONTEXT_ID]: {
           persistence: 'route',
           defaultValue: 'en',
-          values: ['en', 'de', 'ja', 'zh']
+          values: [
+            'en',
+            'de',
+            'ja',
+            'zh',
+            'ru',
+            'fr',
+            'tr',
+            'it',
+            'es',
+            'uk',
+            'pl',
+            'nl',
+            'hi',
+            'ar',
+            'pt',
+            'bn',
+            'pa'
+          ]
         },
         [CURRENCY_CONTEXT_ID]: {
           persistence: 'route',
           defaultValue: 'USD',
-          values: ['USD', 'JPY']
+          values: [
+            'USD',
+            'EUR',
+            'JPY',
+            'GBP',
+            'AUD',
+            'CAD',
+            'CHF',
+            'CNY',
+            'SEK',
+            'NZD',
+            'MXN',
+            'SGD',
+            'HKD',
+            'NOK',
+            'KRW',
+            'TRY',
+            'RUB',
+            'INR',
+            'BRL',
+            'ZAR'
+          ]
         }
-      },
-      urlEncodingParameters: [LANGUAGE_CONTEXT_ID, CURRENCY_CONTEXT_ID]
+      }
     }
   };
 }
