@@ -1,16 +1,14 @@
 import { TestBed } from '@angular/core/testing';
-
 import { provideMockActions } from '@ngrx/effects/testing';
-
 import { Observable, of } from 'rxjs';
-
 import { cold, hot } from 'jasmine-marbles';
 
 import * as fromUserAddressesAction from '../actions/user-addresses.action';
-import { AddressList, Address } from '../../../occ';
-import { OccUserService } from '../../occ/index';
-
 import * as fromUserAddressesEffect from './user-addresses.effect';
+import { AddressList, Address, User } from '../../../occ/index';
+import { OccUserService } from '../../occ/index';
+import { UserService } from '../../facade/user.service';
+import { GlobalMessageService } from '../../../global-message/index';
 
 class MockOccUserService {
   loadUserAddresses(_userId: string): Observable<any> {
@@ -25,6 +23,18 @@ class MockOccUserService {
   deleteUserAddress(): Observable<any> {
     return of({});
   }
+}
+
+class MockUserService {
+  loadAddresses = jasmine.createSpy();
+
+  get(): Observable<User> {
+    return of({});
+  }
+}
+
+class MockGlobalMessageService {
+  add = jasmine.createSpy();
 }
 
 const mockUserAddresses: AddressList = { addresses: [{ id: 'address123' }] };
@@ -42,7 +52,7 @@ const mockUserAddress: Address = {
 
 describe('User Addresses effect', () => {
   let userAddressesEffect: fromUserAddressesEffect.UserAddressesEffects;
-  let userService: OccUserService;
+  let userOccService: OccUserService;
   let actions$: Observable<any>;
 
   beforeEach(() => {
@@ -50,6 +60,8 @@ describe('User Addresses effect', () => {
       providers: [
         fromUserAddressesEffect.UserAddressesEffects,
         { provide: OccUserService, useClass: MockOccUserService },
+        { provide: UserService, useClass: MockUserService },
+        { provide: GlobalMessageService, useClass: MockGlobalMessageService },
         provideMockActions(() => actions$)
       ]
     });
@@ -57,9 +69,9 @@ describe('User Addresses effect', () => {
     userAddressesEffect = TestBed.get(
       fromUserAddressesEffect.UserAddressesEffects
     );
-    userService = TestBed.get(OccUserService);
+    userOccService = TestBed.get(OccUserService);
 
-    spyOn(userService, 'loadUserAddresses').and.returnValue(
+    spyOn(userOccService, 'loadUserAddresses').and.returnValue(
       of(mockUserAddresses)
     );
   });
