@@ -1,5 +1,4 @@
 import { TestBed } from '@angular/core/testing';
-
 import { Store, StoreModule, select } from '@ngrx/store';
 
 import { StateWithUser, USER_FEATURE } from '../user-state';
@@ -60,8 +59,8 @@ describe('User Addresses Selectors', () => {
     });
   });
 
-  describe('getAddressLoading', () => {
-    it('should return isLoading flag', () => {
+  describe('getAddressesLoading', () => {
+    it('should return loading flag', () => {
       let result: boolean;
       store
         .pipe(select(fromSelectors.getAddressesLoading))
@@ -70,21 +69,6 @@ describe('User Addresses Selectors', () => {
       expect(result).toEqual(false);
 
       store.dispatch(new fromActions.LoadUserAddresses('userId'));
-
-      expect(result).toEqual(true);
-    });
-  });
-
-  describe('getAddressProcessingStatus', () => {
-    it('should return isActionProcessing flag', () => {
-      let result: boolean;
-      store
-        .pipe(select(fromSelectors.getAddressesLoading))
-        .subscribe(value => (result = value));
-
-      expect(result).toEqual(false);
-
-      store.dispatch(new fromActions.DeleteUserAddress('addressId'));
 
       expect(result).toEqual(true);
     });
