@@ -1,25 +1,18 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { DebugElement } from '@angular/core';
-import { By } from '@angular/platform-browser';
-
-import { StoreModule } from '@ngrx/store';
-
 import {
-  UserService,
-  CheckoutService,
-  Address,
-  AddressValidation,
-  User
-} from '@spartacus/core';
-import { GlobalMessageService } from '@spartacus/core';
+  DebugElement,
+  Component,
+  Input,
+  Output,
+  EventEmitter
+} from '@angular/core';
+import { By } from '@angular/platform-browser';
+import { BehaviorSubject, of, Observable } from 'rxjs';
 
-import { BehaviorSubject, of } from 'rxjs';
-
+import { Address, User } from '@spartacus/core';
 import { SpinnerModule } from '../../ui/components/spinner/spinner.module';
-import { AddressFormModule } from '../../checkout/components/multi-step-checkout/shipping-address/address-form/address-form.module';
-
 import { AddressBookComponent } from './address-book.component';
-import { AddressBookModule } from './address-book.module';
+import { AddressBookComponentService } from './address-book.component.service';
 
 const mockAddress: Address = {
   id: '123',
@@ -35,59 +28,90 @@ const mockAddress: Address = {
   defaultAddress: false
 };
 
-class MockCheckoutService {
-  clearAddressVerificationResults = jasmine.createSpy();
+const mockUser: User = {
+  uid: '1234'
+};
 
-  verifyAddress = jasmine.createSpy();
+const isLoading = new BehaviorSubject<boolean>(false);
 
-  getAddressVerificationResults(): BehaviorSubject<AddressValidation> {
-    return new BehaviorSubject({ decision: 'ACCEPT' });
+class MockComponentService {
+  loadAddresses = jasmine.createSpy();
+  addUserAddress = jasmine.createSpy();
+  updateUserAddress = jasmine.createSpy();
+  getAddressesStateLoading(): Observable<boolean> {
+    return isLoading.asObservable();
   }
+  getAddresses(): Observable<Address[]> {
+    return of([mockAddress, mockAddress, mockAddress]);
+  }
+  getUserId(): Observable<string> {
+    return of(mockUser.uid);
+  }
+}
+
+@Component({
+  selector: 'cx-address-card',
+  template: ''
+})
+class MockAddressCardComponent {
+  @Input()
+  userId: string;
+
+  @Input()
+  address: Address;
+
+  @Output()
+  editEvent = new EventEmitter<any>();
+}
+
+@Component({
+  selector: 'cx-address-form',
+  template: ''
+})
+class MockAddressFormComponent {
+  @Input()
+  addressData: Address;
+
+  @Input()
+  actionBtnLabel: string;
+
+  @Input()
+  cancelBtnLabel: string;
+
+  @Input()
+  setAsDefaultField: boolean;
+
+  @Input()
+  showTitleCode: boolean;
+
+  @Output()
+  submitAddress = new EventEmitter<any>();
+
+  @Output()
+  backToAddress = new EventEmitter<any>();
 }
 
 describe('AddressBookComponent', () => {
   let component: AddressBookComponent;
   let fixture: ComponentFixture<AddressBookComponent>;
-  let mockUserService: any;
   let el: DebugElement;
-  let mockGlobalMessageService: any;
-  const addresses = new BehaviorSubject<Address[]>([mockAddress]);
-  const isLoading = new BehaviorSubject<boolean>(false);
-  const user = new BehaviorSubject<User>({ uid: 'userId' });
+
+  const checkSectionMsg = (msg: string) => {
+    expect(
+      el.query(By.css('.cx-section__msg')).nativeElement.textContent
+    ).toContain(msg);
+  };
 
   beforeEach(async(() => {
-    mockUserService = {
-      getAddresses: jasmine.createSpy().and.returnValue(addresses),
-      getAddressesLoading: jasmine.createSpy().and.returnValue(isLoading),
-      get: jasmine.createSpy().and.returnValue(user),
-      addUserAddress: jasmine.createSpy(),
-      loadAddresses: jasmine.createSpy(),
-      deleteUserAddress: jasmine.createSpy(),
-      updateUserAddress: jasmine.createSpy(),
-      setAddressAsDefault: jasmine.createSpy(),
-      getDeliveryCountries: jasmine.createSpy().and.returnValue(of([])),
-      getTitles: jasmine.createSpy().and.returnValue(of([])),
-      getRegions: jasmine.createSpy().and.returnValue(of([])),
-      loadTitles: jasmine.createSpy(),
-      loadDeliveryCountries: jasmine.createSpy(),
-      loadRegions: jasmine.createSpy()
-    };
-
-    mockGlobalMessageService = {
-      add: jasmine.createSpy()
-    };
-
     TestBed.configureTestingModule({
-      imports: [
-        AddressBookModule,
-        AddressFormModule,
-        SpinnerModule,
-        StoreModule.forRoot({})
-      ],
+      imports: [SpinnerModule],
       providers: [
-        { provide: UserService, useValue: mockUserService },
-        { provide: CheckoutService, useClass: MockCheckoutService },
-        { provide: GlobalMessageService, useValue: mockGlobalMessageService }
+        { provide: AddressBookComponentService, useClass: MockComponentService }
+      ],
+      declarations: [
+        AddressBookComponent,
+        MockAddressCardComponent,
+        MockAddressFormComponent
       ]
     }).compileComponents();
   }));
@@ -96,6 +120,10 @@ describe('AddressBookComponent', () => {
     fixture = TestBed.createComponent(AddressBookComponent);
     component = fixture.componentInstance;
     el = fixture.debugElement;
+
+    isLoading.next(false);
+    component.ngOnInit();
+    fixture.detectChanges();
   });
 
   it('should create', () => {
@@ -103,129 +131,61 @@ describe('AddressBookComponent', () => {
   });
 
   it('should show spinner if addresses are loading', () => {
-    component.ngOnInit();
     isLoading.next(true);
     fixture.detectChanges();
     expect(el.query(By.css('cx-spinner'))).toBeTruthy();
   });
 
-  it('should show spinner if any action is processing', () => {
-    component.ngOnInit();
-    fixture.detectChanges();
-    expect(el.query(By.css('cx-spinner'))).toBeTruthy();
-  });
-
   it('should show address cards after loading', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress, mockAddress]);
-    fixture.detectChanges();
     expect(el.query(By.css('cx-address-card'))).toBeTruthy();
   });
 
   it('should address cards number to be equal with addresses count', () => {
-    component.ngOnInit();
-    addresses.next([mockAddress, mockAddress]);
-    fixture.detectChanges();
-
-    expect(el.queryAll(By.css('cx-address-card')).length).toEqual(2);
+    expect(el.queryAll(By.css('cx-address-card')).length).toEqual(3);
   });
 
-  it('should show confirmation on delete', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress]);
-    fixture.detectChanges();
-    el.query(By.css('.cx-address-card__actions .delete')).nativeElement.click();
-    fixture.detectChanges();
-    expect(
-      el.query(By.css('.cx-address-card__delete-msg')).nativeElement.textContent
-    ).toContain('Are you sure you want to delete this address?');
-  });
-
-  it('should show adding address form', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress]);
-    fixture.detectChanges();
+  it('should show and hide add address form', () => {
     el.query(By.css('.btn-action')).nativeElement.click();
     fixture.detectChanges();
-    expect(
-      el.query(By.css('cx-address-form')).nativeElement.textContent
-    ).toContain('Add address');
-    expect(component.isAddAddressFormOpen).toBeTruthy();
+    expect(component.showAddAddressForm).toBe(true);
+    checkSectionMsg('Add');
+
+    fixture.whenStable().then(() => {
+      el.query(
+        By.css('.cx-address-form__btns .btn-action')
+      ).nativeElement.click();
+      fixture.detectChanges();
+      checkSectionMsg('Saved');
+    });
   });
 
-  it('should hide adding address form', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress]);
-    fixture.detectChanges();
-    el.query(By.css('.btn-action')).nativeElement.click();
-    fixture.detectChanges();
-    el.queryAll(
-      By.css('.cx-address-form__btns button')
-    )[0].nativeElement.click();
-    fixture.detectChanges();
-    expect(component.isAddAddressFormOpen).toBeFalsy();
+  it('should show and hide edit address form', () => {
+    fixture.whenStable().then(() => {
+      el.query(By.css('.cx-address-card__actions .edit')).nativeElement.click();
+      fixture.detectChanges();
+      expect(component.showEditAddressForm).toBe(true);
+      checkSectionMsg('Edit');
+
+      el.query(
+        By.css('.cx-address-form__btns .btn-action')
+      ).nativeElement.click();
+      fixture.detectChanges();
+      checkSectionMsg('Saved');
+    });
   });
 
-  it('should show editing address form', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress]);
-    fixture.detectChanges();
-    el.query(By.css('.cx-address-card__actions .edit')).nativeElement.click();
-    fixture.detectChanges();
-    expect(
-      el.query(By.css('cx-address-form')).nativeElement.textContent
-    ).toContain('Update address');
-    expect(component.isEditAddressFormOpen).toBeTruthy();
-  });
-
-  it('should hide editing address form', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress]);
-    fixture.detectChanges();
-    el.query(By.css('.cx-address-card__actions .edit')).nativeElement.click();
-    fixture.detectChanges();
-    el.queryAll(
-      By.css('.cx-address-form__btns button')
-    )[0].nativeElement.click();
-    fixture.detectChanges();
-    expect(component.isEditAddressFormOpen).toBeFalsy();
-  });
-
-  it('should successfully set address as default', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress]);
-    fixture.detectChanges();
-    el.query(
-      By.css('.cx-address-card__actions .set-default')
-    ).nativeElement.click();
-    fixture.detectChanges();
-    expect(mockUserService.setAddressAsDefault).toHaveBeenCalledWith(
-      'userId',
-      mockAddress.id
-    );
-  });
-
-  it('should successfully delete address', () => {
-    component.ngOnInit();
-    isLoading.next(false);
-    addresses.next([mockAddress]);
-    fixture.detectChanges();
-    el.query(By.css('.cx-address-card__actions .delete')).nativeElement.click();
-    fixture.detectChanges();
-    el.query(
-      By.css('.cx-address-card__delete .btn-primary')
-    ).nativeElement.click();
-    fixture.detectChanges();
-    expect(mockUserService.deleteUserAddress).toHaveBeenCalledWith(
-      'userId',
-      mockAddress.id
-    );
+  it('should delete address (with confirmation question)', () => {
+    fixture.whenStable().then(() => {
+      expect(el.queryAll(By.css('cx-address-card')).length).toEqual(3);
+      el.query(
+        By.css('.cx-address-card__actions .delete')
+      ).nativeElement.click();
+      fixture.detectChanges();
+      el.query(
+        By.css('.cx-address-card__delete .btn-primary')
+      ).nativeElement.click();
+      fixture.detectChanges();
+      expect(el.queryAll(By.css('cx-address-card')).length).toEqual(2);
+    });
   });
 });
