@@ -1,9 +1,8 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule } from '@angular/router';
+import { ConfigModule, CmsConfig, CmsPageTitleModule } from '@spartacus/core';
 import { BreadcrumbComponent } from './breadcrumb.component';
-import { ConfigModule } from '@spartacus/core';
-import { CmsConfig } from '@spartacus/core';
 
 @NgModule({
   imports: [
@@ -13,10 +12,10 @@ import { CmsConfig } from '@spartacus/core';
       cmsComponents: {
         BreadcrumbComponent: { selector: 'cx-breadcrumb' }
       }
-    })
+    }),
+    CmsPageTitleModule
   ],
   declarations: [BreadcrumbComponent],
-  entryComponents: [BreadcrumbComponent],
-  exports: [BreadcrumbComponent]
+  entryComponents: [BreadcrumbComponent]
 })
 export class BreadcrumbModule {}
