@@ -8,7 +8,7 @@ export * from './navigation/navigation.component';
 export * from './category-navigation/category-navigation.component';
 export * from './product-references/product-references.component';
 export * from './link/link.component';
-export * from './search-box/search-box.component';
+
 export * from './breadcrumb/breadcrumb.component';
 export * from './banner/banner.component';
 export * from './banner/responsive-banner.component';
@@ -21,7 +21,11 @@ export * from './navigation/navigation.module';
 export * from './category-navigation/category-navigation.module';
 export * from './product-references/product-references.module';
 export * from './link/link.module';
+
 export * from './search-box/search-box.module';
+export * from './search-box/search-box.component';
+export * from './search-box/search-box-component.service';
+
 export * from './breadcrumb/breadcrumb.module';
 export * from './banner/banner.module';
 
