@@ -6,7 +6,12 @@ import { RouterModule } from '@angular/router';
 import { MiniCartComponent } from './mini-cart.component';
 
 import { BannerModule } from '../banner/banner.module';
-import { ConfigModule, CmsConfig, UrlTranslationModule } from '@spartacus/core';
+import {
+  ConfigModule,
+  CmsConfig,
+  UrlTranslationModule,
+  CartModule
+} from '@spartacus/core';
 
 @NgModule({
   imports: [
@@ -14,6 +19,7 @@ import { ConfigModule, CmsConfig, UrlTranslationModule } from '@spartacus/core';
     RouterModule,
     BannerModule,
     MediaModule,
+    CartModule,
     ConfigModule.withConfig(<CmsConfig>{
       cmsComponents: {
         MiniCartComponent: { selector: 'cx-mini-cart' }
