@@ -1,7 +1,8 @@
-import { DebugElement, Pipe, PipeTransform } from '@angular/core';
+import { Pipe, PipeTransform, DebugElement } from '@angular/core';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-import { By } from '@angular/platform-browser';
 import { RouterTestingModule } from '@angular/router/testing';
+import { of, Observable } from 'rxjs';
+import { By } from '@angular/platform-browser';
 
 import {
   ProductService,
@@ -9,11 +10,8 @@ import {
   Component,
   CmsProductCarouselComponent
 } from '@spartacus/core';
-
-import { of, Observable } from 'rxjs';
-
+import { ProductCarouselService } from './product-carousel.component.service';
 import { PictureComponent } from '../../ui/components/media/picture/picture.component';
-
 import { ProductCarouselComponent } from './product-carousel.component';
 import { CmsComponentData } from '../../cms/components/cms-component-data';
 
@@ -60,6 +58,35 @@ class MockProductService {
   }
 }
 
+class MockProductCarouselService {
+  getTitle = jasmine.createSpy('getTitle').and.callFake(() => of('Mock Title'));
+  setTitle = jasmine.createSpy('setTitle').and.callFake(() => of('Mock Title'));
+  getItems = jasmine
+    .createSpy('getItems')
+    .and.callFake(() =>
+      of([of(mockProduct), of(mockProduct), of(mockProduct), of(mockProduct)])
+    );
+  setItems = jasmine
+    .createSpy('setItems')
+    .and.callFake(() =>
+      of([of(mockProduct), of(mockProduct), of(mockProduct), of(mockProduct)])
+    );
+
+  getItemSize = jasmine.createSpy('getItemSize').and.callFake(() => of(4));
+  setItemSize = jasmine.createSpy('setItemSize');
+  getItemAsActive = jasmine
+    .createSpy('getItemAsActive')
+    .and.callFake(() => of(0));
+  setItemAsActive = jasmine
+    .createSpy('setItemAsActive')
+    .and.callFake(() => of(1));
+  setPreviousItemAsActive = jasmine.createSpy('setPreviousItemAsActive');
+  getActiveItemWithDelay = jasmine.createSpy('getActiveItemWithDelay');
+  setNextItemAsActive = jasmine.createSpy('setNextItemAsActive');
+  getDelayValue = jasmine.createSpy('getDelayValue').and.callThrough();
+  getActiveItem = jasmine.createSpy('getActiveItem').and.callFake(() => of(1));
+}
+
 describe('ProductCarouselComponent', () => {
   let productCarouselComponent: ProductCarouselComponent;
   let fixture: ComponentFixture<ProductCarouselComponent>;
@@ -77,7 +104,18 @@ describe('ProductCarouselComponent', () => {
         { provide: CmsComponentData, useValue: MockCmsComponentData },
         { provide: ProductService, useClass: MockProductService }
       ]
-    }).compileComponents();
+    })
+      .overrideComponent(ProductCarouselComponent, {
+        set: {
+          providers: [
+            {
+              provide: ProductCarouselService,
+              useClass: MockProductCarouselService
+            }
+          ]
+        }
+      })
+      .compileComponents();
   }));
 
   beforeEach(() => {
@@ -87,33 +125,22 @@ describe('ProductCarouselComponent', () => {
     el = fixture.debugElement;
   });
 
-  it('should be created', () => {
+  it('should be created', async(() => {
+    productCarouselComponent.ngOnInit();
     expect(productCarouselComponent).toBeTruthy();
-  });
+  }));
 
-  it('should have products', () => {
-    expect(productCarouselComponent.items$);
-
+  it('should have products', async(() => {
     let products$: Observable<Product>[];
-    productCarouselComponent.items$
+    productCarouselComponent.service.setItems();
+    productCarouselComponent.service
+      .getItems()
       .subscribe(productData$ => {
         products$ = productData$;
       })
       .unsubscribe();
     expect(products$.length).toBe(productCodeArray.length);
-  });
-
-  it('should have product data', () => {
-    expect(productCarouselComponent.items$);
-
-    let product: Product;
-    productCarouselComponent.items$
-      .subscribe(productData$ => {
-        productData$[0].subscribe(data => (product = data)).unsubscribe();
-      })
-      .unsubscribe();
-    expect(product).toBe(mockProduct);
-  });
+  }));
 
   it('should contain cms content in the html rendering after bootstrap', () => {
     expect(el.query(By.css('h3')).nativeElement.textContent).toContain(
