@@ -1,5 +1,5 @@
 import { Injectable } from '@angular/core';
-import { Observable, of } from 'rxjs';
+import { Observable } from 'rxjs';
 import { switchMap, distinctUntilChanged, map, filter } from 'rxjs/operators';
 import { CmsService, Page } from '@spartacus/core';
 import { BreakpointService } from '../../ui/layout/breakpoint/breakpoint.service';
@@ -42,36 +42,8 @@ export class PageLayoutService {
     );
   }
 
-  private showTitle(section?: string): Observable<boolean> {
-    return this.breakpointService.breakpoint$.pipe(
-      switchMap(breakpoint =>
-        this.page$.pipe(
-          map(page => {
-            return this.getSlotConfig(
-              page.template,
-              'showTitle',
-              section,
-              breakpoint
-            );
-          }),
-          filter(Boolean),
-          map(config => config.showTitle)
-        )
-      ),
-      distinctUntilChanged()
-    );
-  }
-
   get page$(): Observable<Page> {
     return this.cms.getCurrentPage().pipe(filter(Boolean));
-  }
-
-  get pageTitle$(): Observable<string> {
-    return this.showTitle().pipe(
-      switchMap(show =>
-        show ? this.page$.pipe(map((page: Page) => page.title)) : of()
-      )
-    );
   }
 
   get templateName$(): Observable<string> {
