@@ -1,31 +1,6 @@
 export * from './ui.module';
 export * from './ui-framework/ui-framework.module';
-export * from './layout/layout.module';
-export * from './components/components.module';
-export * from './components/media/media.module';
-export * from './components/form-components/form-components.module';
-export * from './components/pagination-and-sorting/pagination-and-sorting.module';
-export * from './components/spinner/spinner.component';
-export * from './pages/pages.module';
-export * from './layout/order-confirmation-page-layout/order-confirmation-page-layout.component';
-export * from './layout/main/storefront.component';
-export * from './layout/header/header.component';
-export * from './layout/store-finder-page-layout/store-finder-page-layout.component';
-export * from './components/media/picture/picture.component';
-export * from './components/form-components/star-rating/star-rating.component';
-export * from './components/form-components/item-counter/item-counter.component';
-export * from './components/generic-link/generic-link.component';
-export * from './pages/category-page/category-page.component';
-export * from './pages/product-page/product-page.component';
-export * from './pages/store-finder-page/store-finder-page.component';
-export * from './pages/cart-page/cart-page.component';
-export * from './pages/order-confirmation-page/order-confirmation-page.component';
-export * from './pages/login-page/login-page.component';
-export * from './pages/cart-page/cart-page.module';
-export * from './pages/category-page/category-page.module';
-export * from './pages/product-page/product-page.module';
-export * from './pages/register-page/register-page.module';
-export * from './pages/login-page/login-page.module';
-export * from './pages/store-finder-page/store-finder-page.module';
 
-export * from './layout/main/main.module';
+export * from './components/index';
+export * from './pages/index';
+export * from './layout/index';
