@@ -42,7 +42,6 @@ export const defaultLayoutConfig: LayoutConfig = {
     ProductDetailsPageTemplate: {
       slots: [
         'TopHeaderSlot',
-        'BottomHeaderSlot',
         'VariantSelectorSlot',
         'AddToCart',
         'UpSelling',
@@ -59,18 +58,19 @@ export const defaultLayoutConfig: LayoutConfig = {
       ]
     },
     AccountPageTemplate: {
-      showTitle: true,
       slots: ['BodyContent', 'SideContent']
     },
     LoginPageTemplate: {
-      showTitle: true,
       slots: ['LeftContentSlot', 'RightContentSlot']
     },
     ErrorPageTemplate: {
       slots: ['MiddleContent']
     },
+    OrderConfirmationPageTemplate: {
+      slots: ['BodyContent', 'SideContent']
+    },
     MultiStepCheckoutSummaryPageTemplate: {
-      slots: ['SideContent', 'BodyContent']
+      slots: ['BodyContent', 'SideContent']
     }
   }
 };
