@@ -1,20 +1,18 @@
 import { NgModule } from '@angular/core';
 
+import { ConfigModule, Config } from '@spartacus/core';
+
 import { OutletRefModule } from '../../outlet/index';
 import { StyleRefModule } from '../../outlet/style-ref/style-ref.module';
 
-import { MainModule } from './main/main.module';
-import { OrderConfirmationPageLayoutModule } from './order-confirmation-page-layout/order-confirmation-page-layout.module';
-import { RegisterLayoutModule } from './register-layout/register-layout.module';
-import { StoreFinderPageLayoutModule } from './store-finder-page-layout/store-finder-page-layout.module';
 import { BreakpointService } from './breakpoint/breakpoint.service';
-import { ConfigModule, Config } from '@spartacus/core';
-
 import { defaultLayoutConfig } from './config/default-layout-config';
 import { LayoutConfig } from './config/layout-config';
+import { MainModule } from './main/main.module';
+import { RegisterLayoutModule } from './register-layout/register-layout.module';
+import { StoreFinderPageLayoutModule } from './store-finder-page-layout/store-finder-page-layout.module';
 
 const layoutModules = [
-  OrderConfirmationPageLayoutModule,
   RegisterLayoutModule,
   StoreFinderPageLayoutModule,
   OutletRefModule,
