@@ -1,8 +1,9 @@
 import { Component } from '@angular/core';
-
 @Component({
   selector: 'cx-storefront',
   templateUrl: './storefront.component.html',
   styleUrls: ['./storefront.component.scss']
 })
-export class StorefrontComponent {}
+export class StorefrontComponent {
+  constructor() {}
+}
