@@ -2,13 +2,18 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
 
-import { CmsPageGuards } from '../../../cms/guards/cms-page.guard';
-import { OrderConfirmationPageLayoutModule } from '../../layout/order-confirmation-page-layout/order-confirmation-page-layout.module';
-import { OrderConfirmationPageComponent } from './order-confirmation-page.component';
-import { OrderConfirmationPageGuard } from '../../../checkout/guards/order-confirmation-page.guard';
 import { AuthGuard } from '@spartacus/core';
 
+import { OrderConfirmationModule } from '../../../checkout/index';
+import { PageLayoutModule } from '../../../cms/index';
+import { OutletRefModule } from '../../../outlet/index';
+import { OrderConfirmationPageGuard } from '../../../checkout/guards/order-confirmation-page.guard';
+import { CmsPageGuards } from '../../../cms/guards/cms-page.guard';
+
+import { OrderConfirmationPageComponent } from './order-confirmation-page.component';
+
 const routes: Routes = [
+  // TODO: as soon as the components are moved to CMS driven components we can drop this specific OrderConfirmationPageComponent
   {
     path: null,
     canActivate: [AuthGuard, CmsPageGuards, OrderConfirmationPageGuard],
@@ -20,9 +25,12 @@ const routes: Routes = [
 @NgModule({
   imports: [
     CommonModule,
-    OrderConfirmationPageLayoutModule,
+    OrderConfirmationModule,
+    PageLayoutModule,
+    OutletRefModule,
     RouterModule.forChild(routes)
   ],
+  providers: [OrderConfirmationPageGuard],
   declarations: [OrderConfirmationPageComponent],
   exports: [OrderConfirmationPageComponent]
 })
